backend/requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
starlette==0.38.6

# Configuration management
PyYAML==6.0.2

# Database and vector storage
chromadb==0.5.15
sentence-transformers==3.1.1

# Data validation and parsing
pydantic==2.9.2
pydantic_core==2.27.1
annotated-types==0.7.0

# HTTP and networking
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10

# Docker integration
docker==7.1.0

# Git integration
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# WebSocket support
websockets==13.1
python-websockets==10.4

# Async support
anyio==4.6.0
sniffio==1.3.1

# Utilities
python-dotenv==1.0.1
click==8.1.7
h11==0.14.0
typing_extensions==4.12.2

# Slack integration (for slack connector)
slack-bolt==1.21.2
slack-sdk==3.33.2

# LLM integration
openai==1.51.2
anthropic==0.37.1

# Additional ML/AI dependencies
numpy==2.1.2
torch==2.5.1
transformers==4.46.2
huggingface-hub==0.26.2

# Development and testing
pytest==8.3.3
pytest-asyncio==0.24.0
